# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DedicationInputTone = typeof DedicationInputTone[keyof typeof DedicationInputTone];


export const DedicationInputTone = {
  romantic: 'romantic',
  leve: 'leve',
  poetic: 'poetic',
} as const;

export type DedicationInputProvider = typeof DedicationInputProvider[keyof typeof DedicationInputProvider];


export const DedicationInputProvider = {
  auto: 'auto',
  groq: 'groq',
  gemini: 'gemini',
  mistral: 'mistral',
} as const;

export interface DedicationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  recipientName: string;
  /** @maxLength 80 */
  senderName?: string;
  tone: DedicationInputTone;
  provider?: DedicationInputProvider;
  /** @maxLength 600 */
  details?: string;
  /**
     * @maxItems 12
     * @items.minLength 1
     * @items.maxLength 180
     */
  memories?: string[];
}

export interface Dedication {
  text: string;
  title: string;
}

export interface Error {
  error: string;
}