# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: AI-powered dedication generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openai/dedication:
    post:
      operationId: generateDedication
      tags: [openai]
      summary: Generate a personalized dedication
      description: Creates a warm, respectful romantic dedication based on the recipient's interests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DedicationInput"
      responses:
        "200":
          description: Generated dedication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dedication"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DedicationInput:
      type: object
      properties:
        recipientName:
          type: string
          minLength: 1
          maxLength: 80
        senderName:
          type: string
          maxLength: 80
        tone:
          type: string
          enum: [romantic, leve, poetic]
        provider:
          type: string
          enum: [auto, groq, gemini, mistral]
          default: auto
        details:
          type: string
          maxLength: 600
        memories:
          type: array
          maxItems: 12
          items:
            type: string
            minLength: 1
            maxLength: 180
      required:
        - recipientName
        - tone
    Dedication:
      type: object
      properties:
        text:
          type: string
        title:
          type: string
      required:
        - text
        - title
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error