# lib/api-zod/src/generated/types/dedicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DedicationInputProvider } from './dedicationInputProvider';
import type { DedicationInputTone } from './dedicationInputTone';

export interface DedicationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  recipientName: string;
  /** @maxLength 80 */
  senderName?: string;
  tone: DedicationInputTone;
  provider?: DedicationInputProvider;
  /** @maxLength 600 */
  details?: string;
  /**
     * @maxItems 12
     * @items.minLength 1
     * @items.maxLength 180
     */
  memories?: string[];
}